requirements.txt
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
coverage==7.10.6
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hypothesis==6.138.14
idna==3.10
iniconfig==2.1.0
mako==1.3.10
markupsafe==3.0.2
mypy==1.17.1
mypy-extensions==1.1.0
orjson==3.11.3
packaging==25.0
pathspec==0.12.1
pip==25.2
pluggy==1.6.0
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pynacl==1.5.0
pytest==8.4.2
pytest-asyncio==1.1.0
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
ruff==0.12.12
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.43
starlette==0.47.3
structlog==25.4.0
types-requests==2.32.4.20250809
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
